Minor change
Added a seek directory to standard read (always from the beginning of
the stream), to make sure the code behaves as expected.
I don't know how this will impact performance yet, might change this
later on, if performance takes a hit because of this, but for now I tink
it's easier that way.