Yet another change to how I handle frame data
I'm now back to just reading it as string data, and just using the
data() function of the std::string class to get a char* if I need the
raw bytes.

This is probably the simplest solution, so that's what I'll use for now.

Also finished implementation of frames that rely on the raw bytes.