Fixed issue with reading PCNT frame data
I changed readFrame again...

[Changes]:
The data is put into a vector and a shared pointer to that vector is
then returned.

I also wrote a small function to null terminate that data if neccessary.

[Reasons]
Turns out just treating everything as a string is not such a great idea
after all, since the string will have no content if the first byte of
the data is 0x00.
That will then be treated as the null terminator for the string, so data
will be lost.
This happens fairly often with PCNT frames, so in order to read those
correctly (and avoid bugs in the future), I had to change it again.