Changed decoding of text frames
The first byte of a text frame is the encoding method, the sizes or
anything are not messed up. So instead of just converting the data to a
string right away, it is first decoded.

So the sizes and position pointers when reading frame data is no longer
an issue.