Added crude error handling to decode_text function(s)
Added an error flag to the struct that holds the decoded text and the
position that will be set when the encoding method supplied to the
function is not valid.

However the decode_text function that used to only return the string now
returns a nullptr that I have to check for in case of an error (and
logs the error message). I don't know if that's a good idea yet or if I
want to keep it that way, but better than nothing at all I suppose.