Fixed convert_bytes function
I for some reason uncluded syncsafety into the convert_bytes function,
which doesn't make sense at all (I'm converting from bytes to an
integer, not the other way around), and that was sometimes screwing with
tag/frame sizes. That is fixed now.

I also adjusted the documentation and fixed a comment.

And I removed the constexpr declaration. I don't know why I don't get an
error when it is there, because that function depends on data that is
fed to it after reading the file, which is not possible at compiletime.
I'm a little confused by what is going on there, but unless I'm wrong,
constexpr should not be there.