Fixed convert_size function
Fixed function that converts an (unsigned) 32 bit integer into 4
(unsigned) bytes, that then can be written back into the mp3 file as a
tag/frame size.

The older version got the conversion completely wrong, so the conversion
should be correct now, and it completely disregarded syncsafety, which I
now check as well. However I still kind of have to deal with the
occasion if that happens, but I don't know how yet.