Fixed bug with size calculation
Fixed a bug where the size would be wrong due to bad conversion.
When converting from char to std::uint64_t the data would not be filled
with 0x00 bytes, but 0xff bytes if the char had a negative value.

That would fuck up conversions a lot. Unfortunately I can't read
'unsigned bytes' from a file, so I have to convert it all before
converting.